fix som bugs

+++ scheduler.c
@@ -10,7 +10,8 @@
 
 
 FILE* logFile, *perfFile;
-ALGORITHM algorithm;
+int algorithm;
+int algo;
 PriorityQueue readyQ;
 Process* Process_Table;
 Process* running = NULL;
@@ -21,9 +22,12 @@ int* shmRemainingtime;
 int current_process_id;
 int total_number_of_received_process;
 int total_number_of_processes;
+int RR_Priority;
+bool if_termination;
 
 bool process_generator_finished = false;
 
+int clk;
 
 
 int sem;
@@ -33,9 +37,9 @@ union Semun semun;
 union Semun
 {
     int val;               /* value for SETVAL */
-    struct semid_ds *buf;  /* buffer for IPC_STAT & IPC_SET */
-    ushort *array;         /* array for GETALL & SETALL */
-    struct seminfo *__buf; /* buffer for IPC_INFO */
+    struct semid_ds buf;  /* buffer for IPC_STAT & IPC_SET */
+    ushort array;         /* array for GETALL & SETALL */
+    struct seminfo __buf; /* buffer for IPC_INFO */
     void *__pad;
 };
 
@@ -122,7 +126,7 @@ int main(int argc, char * argv[])
 
    
 
-    //signal(SIGUSR1, handler_notify_scheduler_new_process_has_arrived);
+    signal(SIGUSR1, handler_notify_scheduler_new_process_has_arrived);
     signal(SIGUSR2, ProcessTerminates);
  
 
@@ -147,14 +151,14 @@ int main(int argc, char * argv[])
 
 
     //semaphore
-    key_id = ftok("key", 55);
-    sem = semget(key_id, 1, 0666 | IPC_CREAT);
-    semun.val = 0; /* initial value of the semaphore, Binary semaphore */
-    if (semctl(sem, 0, SETVAL, semun) == -1)
-    {
-        perror("Error in semctl");
-        exit(-1);
-    }
+    // key_id = ftok("key", 55);
+    // sem = semget(key_id, 1, 0666 | IPC_CREAT);
+    // semun.val = 0; /* initial value of the semaphore, Binary semaphore */
+    // if (semctl(sem, 0, SETVAL, semun) == -1)
+    // {
+    //     perror("Error in semctl");
+    //     exit(-1);
+    // }
 
 
     /* Create a message buffer between process_generator and scheduler */
@@ -185,46 +189,49 @@ int main(int argc, char * argv[])
     
     if(argc < 3) { perror("Too few CLA!!"); return -1;}
 
-    switch (argv[2][0])
-    {
-    case '1':
-        algorithm = 0;//HPF_ALGORITHM;
-        break;
-    
-    case '2':
-        algorithm = 1;//SRTN_ALGORITHM;
-        break;
-    case '3':
-        algorithm = 2;//RR_ALGORITHM;
+    // switch (argv[2][0])
+    // {
+        
+    algorithm = (argv[2][0] -'0')-1;
+      printf(".............%d\n",algorithm) ; 
+    if(algorithm == 2){
+        //RR_ALGORITHM;
         if(argc < 4) { perror("Too few CLA!!"); return -1;}
         i = 0;
         Q = 0;
         while(argv[3][i])
             Q = Q * 10 + (argv[3][i++] - '0');
         printf("\nquantum: %d\n", Q);
-        break;
-        default:
-        perror("undefined algorithm");
-        return -1;
+        // break;
+        // default:
+        // perror("undefined algorithm");
+        // return -1;
     }
+    algo = algorithm;
+
+
+
 
     logFile = fopen("Scheduler.log", "w");
     fprintf(logFile, "#At  time  x  process  y  state  arr  w  total  z  remain  y  wait  k\n");//should we ingnore this line ?
+    fflush(0);
     
 
-    switch (algorithm)
-    {
-    case RR_ALGORITHM:
-        RR(Q);
-        break;
-    case HPF_ALGORITHM:
-        HPF();
-        break;
+    RR_Priority = 0;
 
-    case SRTN_ALGORITHM:
+    while(pq_isEmpty(&readyQ)); //to guarantee that once the first process arrives it will begin at once not at the next sec
+
+    if(algo== 2)
+        RR(Q);
+        
+    else if(algo== 0)
+        HPF();
+       
+
+    else if(algo== 1)
         SRTN();
-        break;
-    }
+       
+
 
     
     
@@ -248,14 +255,15 @@ void RR(int quantum)
     int clk = getClk();
     //int timeToStop;
     int currentQuantum = quantum;
-    //handler_notify_scheduler_new_process_has_arrived(0);
+    int remain_beg;
+    //printf("----------the quantum %d\n", quantum);
     while(total_number_of_processes+1)
     {
         printf("###############################################################\n");
         printf("Total Number of processes until now: %d\n", pq_getLength(&readyQ));
         printf("###############################################################\n");
 
-        handler_notify_scheduler_new_process_has_arrived(0);
+        //handler_notify_scheduler_new_process_has_arrived(0);
         //printf("\ni am here -------------------------------------\n");
 
         //printf("\ni am here \n");
@@ -267,30 +275,42 @@ void RR(int quantum)
             //(*running).remainingTime = *shmRemainingtime;
             running->cumulativeRunningTime++;
             
-
+            //to avoid stopping the process after ending the quantum if it is the only process in the system
+            if(currentQuantum == 0 && pq_isEmpty(&readyQ))
+            {
+                currentQuantum = quantum;
+                continue;
+            }
+            
             if(currentQuantum == 0)
             {
+                
                 running->state = WAITING;
 
                 //send signal stop to this process and insert it back in the ready queue
                 running->waiting_start_time = getClk();
                 //down(sem);
-
+                //while(running->remainingTime == *shmRemainingtime);
+                while(remain_beg - *shmRemainingtime == quantum);
                 kill(running->pid, SIGSTOP);
 
                 running->remainingTime = *shmRemainingtime;
 
                 //termination occur ??
-                if(running->remainingTime == 0)
-                    continue;
                 
-                pq_push(&readyQ, running, 0);
+                // if(running->remainingTime == 0)
+                //     continue;
+                
+                RR_Priority++;
+                pq_push(&readyQ, running, RR_Priority);
 
                 //termination occur ??
 
                 write_in_logfile_stopped();
                 
                 running = NULL;
+
+                continue;  //to make the next process begin exactly after the current is stopped
                 //printf("\ni am here after blocking a process--------------------------------------\n");
             }
         }
@@ -299,6 +319,7 @@ void RR(int quantum)
             if(pq_peek(&readyQ))
             {
                 //printf("\ni am here -------------------------------------\n");
+                if_termination = false;
                 running = pq_pop(&readyQ);
                 current_process_id = running->id;
                 
@@ -325,18 +346,23 @@ void RR(int quantum)
                     running->state = RUNNING;
                     running->running_start_time = getClk();
                     
-                    currentQuantum--;
+                    remain_beg = running->burstTime;
+                    //currentQuantum--;
                     write_in_logfile_start();
                 }
                 else{
                     //wake it up
-                    *shmRemainingtime = running->remainingTime;
+                    //*shmRemainingtime = running->remainingTime;
+                    printf("---------------------old remaining : %d\n", *shmRemainingtime);
+                    remain_beg = running->remainingTime;
                     kill(running->pid, SIGCONT); //TO ASK
+                    printf("---------------------new remaining : %d\n", *shmRemainingtime);
+                    running->remainingTime == *shmRemainingtime;
                     running->state = RUNNING;
                     running->running_start_time = getClk();
                     //down(sem);
                     //*shmRemainingtime = running->remainingTime;
-                    currentQuantum--;
+                    //currentQuantum--;
                     write_in_logfile_resume();
                 }
             }
@@ -344,16 +370,11 @@ void RR(int quantum)
 
         }
 
-        //printf("\ni am here before update-------------------------------------\n");
+        
         updateInformation();
-        //printf("\ni am here afger updete-------------------------------------\n");
         printf("\nclk = %d   getclk = %d\n", clk, getClk());
-        while(clk == getClk()){
-            //printf("\n i am inside the while\n");
-        }
-        //printf("\ni am outside the while\n");
+        while(clk == getClk());  // && !pq_isEmpty(&readyQ) && running
         clk = getClk();
-        //printf("\ni am here after clk-------------------------------------\n");
 
 
         // semun.val = 0; /* initial value of the semaphore, Binary semaphore */
@@ -369,8 +390,9 @@ void RR(int quantum)
 /* Warning: Under development */
 void HPF(void)
 {
+    printf("in hpf.............%d\n",algorithm) ; 
     int pid;
-    int clk = getClk();
+     clk = getClk();
     int pr;
 
     running = NULL;
@@ -378,7 +400,7 @@ void HPF(void)
     while(total_number_of_processes)
     {
 
-        handler_notify_scheduler_new_process_has_arrived(0);
+        //handler_notify_scheduler_new_process_has_arrived(0);
 
         printf("###############################################################\n");
         printf("ReadyQ length until now: %d\n", pq_getLength(&readyQ));
@@ -422,7 +444,6 @@ void HPF(void)
             }
         }
         updateInformation();
-
         while(clk == getClk());
         clk=getClk();
         //signal(SIGUSR1, handler_notify_scheduler_new_process_has_arrived);
@@ -430,12 +451,14 @@ void HPF(void)
 }
 void SRTN(void)
 {
+    printf("heeeeeeeeeeeeeeeeeeeeeeeere");
+    fflush(0);
     int clk = -1;
     int peek;
     int pid, pr;
     do
     {
-        current_process_id = running->id;
+        //current_process_id = running->id;
             
 
         if(getClk() != clk)
@@ -453,7 +476,7 @@ void SRTN(void)
                     continue;
                 }
                 
-                down(sem);
+                
                 running->remainingTime = *shmRemainingtime;
 
                 peek = pq_peek(&readyQ)->remainingTime;
@@ -468,8 +491,8 @@ void SRTN(void)
 
                 running->state = WAITING;
                 //send signal stop to this process and insert it back in the ready queue
-                running->waiting_start_time = getClk();
-                kill(running->pid, SIGTSTP);
+                running->waiting_start_time = getClk()-1;
+                kill(running->pid, SIGSTOP);
                 pq_push(&readyQ, running, running->remainingTime);
 
                 write_in_logfile_stopped();
@@ -501,7 +524,7 @@ void SRTN(void)
                 }
                 running->state = RUNNING;
                 running->running_start_time = getClk();
-                down(sem);
+                
                 *shmRemainingtime = running->remainingTime;
 
                 running->pid = pid;
@@ -514,11 +537,11 @@ void SRTN(void)
                 kill(running->pid, SIGCONT);
                 running->state = RUNNING;
                 running->running_start_time = getClk();
-                down(sem);
+              
                 *shmRemainingtime = running->remainingTime;
                 write_in_logfile_resume();
             }
-            down(sem);
+           
             current_process_id = running->id;
             updateInformation();
         }
@@ -659,8 +682,11 @@ void ProcessTerminates(int signum)
     //should we check on the total number of processes and if it equals 0 then terminate the scheduler
 
     printf("\n after process terminates-------------------------\n");
-
+    //int dummy=getClk();
+    //while(dummy == getClk());
+    if_termination = true;
     signal(SIGUSR2, ProcessTerminates);
+    
 }
 
 
@@ -696,21 +722,25 @@ void handler_notify_scheduler_new_process_has_arrived(int signum)
         // pq_push(&readyQ, &Process_Table[msgbuf.id], 0);
 
         //signal(SIGUSR1, handler_notify_scheduler_new_process_has_arrived);
-        
-        if (algorithm == 0)
+         // pq_push(&readyQ, &Process_Table[msgbuf.id], Process_Table[msgbuf.id].priority);
+//          printf("\nalgo///////////////////%d\n",algorithm);
+// printf("\nalgo///////////////////%d\n",algo);
+
+        if (algo==0)
             pq_push(&readyQ, &Process_Table[msgbuf.id], Process_Table[msgbuf.id].priority);
-        else if (algorithm == 1) { /* WARNING: This needs change depends on the SRTN algorithm */
+        else if (algo == 1) { /* WARNING: This needs change depends on the SRTN algorithm */
             #if (WARNINGS == 1)
             #warning "Scheduler: You should decide what will be the priority parameter in the priority queue in case of SRTN algorithm."
             #endif
             pq_push(&readyQ, &Process_Table[msgbuf.id], Process_Table[msgbuf.id].remainingTime);
         }
-        else if (algorithm == 2) {
+        else if (algo == 2) {
             #if (WARNINGS == 1)
             #warning "Scheduler: You should decide what will be the priority parameter in the priority queue in case of RR algorithm."
             #endif
             printf("i am pushing here \n");
-            pq_push(&readyQ, &Process_Table[msgbuf.id], 0);
+            RR_Priority++;
+            pq_push(&readyQ, &Process_Table[msgbuf.id], RR_Priority);
         }
 
 
@@ -723,7 +753,7 @@ void handler_notify_scheduler_new_process_has_arrived(int signum)
     }
     
 
-    //signal(SIGUSR1, handler_notify_scheduler_new_process_has_arrived);
+   //signal(SIGUSR1, handler_notify_scheduler_new_process_has_arrived);
 }
